+++ src/ui/common/editors/Reflection.h
@@ -3,10 +3,10 @@
 inline float rflMinFloatStep{ 0.01f };
 inline unsigned int rflSliderCutOff{ 3000 };
 
-bool ReflectionEditor(const char* label, void* reflectionData, const hh::fnd::RflClass* rflClass, bool defaultOpen = true);
+bool ReflectionEditor(const char* label, void* reflectionData, const hh::fnd::RflClass* rflClass, bool defaultOpen = false);
 
 template<typename T, typename = decltype(T::rflClass)>
-static bool Editor(const char* label, T& reflectionData, bool defaultOpen = true) {
+static bool Editor(const char* label, T& reflectionData, bool defaultOpen = false) {
     return ReflectionEditor(label, &reflectionData, &RESOLVE_STATIC_VARIABLE(T::rflClass), defaultOpen);
 }