+++ src/ui/game-services/game-service-inspectors/rangers/TimeService.cpp
@@ -1,0 +1,35 @@
+#include "TimeService.h"
+
+bool Editor(const char* label, app::game::TimeService::TimeProgressSpeed& speed);
+
+#include <ui/common/editors/rangers/TimeService.h>
+#include <ui/common/editors/Basic.h>
+#include <ui/common/viewers/Basic.h>
+
+bool Editor(const char* label, app::game::TimeService::TimeProgressSpeed& speed) {
+    bool edited{};
+
+    if (ImGui::TreeNode(label)) {
+        edited |= Editor("Start time", speed.start);
+        edited |= Editor("Speed", speed.speed);
+        ImGui::TreePop();
+    }
+
+    return edited;
+}
+
+void RenderGameServiceInspector(app::game::TimeService& service) {
+    Editor("Current time", service.currentTime);
+    //Editor("Previous time", service.previousTime);
+
+    Editor("Start of night in hours", service.startOfNightInHours);
+    Editor("End of night in hours", service.endOfNightInHours);
+
+    Editor("Time progress speeds", service.timeProgressSpeeds);
+
+    Viewer("Night times loaded", service.sceneDataLoaded);
+    Editor("Time progress enabled", service.timeProgressEnabled);
+    Editor("Using fixed time", service.fixedTime);
+
+    Viewer("Seconds until next update", service.secondsUntilNextUpdate);
+}